pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.45.1",
]
